Kay_Gee_1.0/requirements.txt
# Core ML and Scientific Computing
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Graph and Network Analysis
networkx>=3.1

# Database
sqlalchemy>=2.0.0

# Cryptography for Handshake Protocol
cryptography>=41.0.0
pynacl>=1.5.0

# Prolog Integration
pyswip>=0.2.10

# Natural Language Processing (Classical, Non-Neural)
nltk>=3.8.0

# Similarity Search
annoy>=1.17.0

# Audio Processing and Speech
torch>=2.0.0
torchaudio>=2.0.0
TTS>=0.22.0
faster-whisper>=0.10.0
librosa>=0.9.2
soundfile>=0.12.1
pydub>=0.25.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
pylint>=2.17.0
mypy>=1.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging
structlog>=23.1.0
